renaming bin directories and files
