+++ cmd/faBin/faBin_test.go
