+++ cmd/faBin/testdata/testGenome.fa
@@ -1,0 +1,8 @@
+>chr1
+ACGTACGT
+>chr2
+ACGTA
+>chr3
+CGT
+>chr4
+AC
