+++ faBin/testdata/testContigs.fa
